refactor(engine): sıralama ve skor terimleri kendi modüllerine ayrıldı (ALG-05)
ItemOrdering.cs (SortForGroupPlacement, ApplyCriteriaSort), BalanceScoring.cs
(BalancePenalty, ImproveBalance, SwapIsValid, GlobalBalancePenalty) ve
VolumeScoring.cs (derinlik/genişlik terimleri) oluşturuldu.

ComputeScore switch'ten terim toplamına dönüştü:
gravity + depth + balance + width + zone. Kapalı modülün terimi 0m sabiti;
ölçeği yüksek bir sıfır (katsayı*0) yerine erken dönüş kullanıldı ki decimal
ölçek düşürme kaynaklı yuvarlama tetiklenmesin. Terim sırası her kriterin
bugünkü ifadesindeki sırayla birebir aynı tutuldu.

Katsayılar (1e6 / 900_000 / 1_000 / 500 / 2_000) birebir korundu.

Davranış sıfır değişim: snapshot'ların hiçbiri değişmedi. Passed 23 + 20.
Performans: VolumeFirst 5381 ms, WeightBalance 11301 ms, Lifo 5214 ms —
taban çizgisinde, regresyon yok.
